requirements.txt
requests
soundfile
numpy
pyaudio
openai-whisper
torch 